+++ tests/test_settings.py
@@ -3,9 +3,8 @@ import datetime
 from django.core.exceptions import ImproperlyConfigured
 from django.test import TestCase, override_settings
 
+import sesame.test  # noqa
 from sesame import settings
-
-from .signals import reset_sesame_settings  # noqa
 
 
 class TestSettings(TestCase):